// apps/web/src/lib/api/generated/main/model/guild-kill-stats-response-dto-output-member-ranking-item.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { GuildKillStatsResponseDtoOutputMemberRankingItemParticipationsByType } from './guild-kill-stats-response-dto-output-member-ranking-item-participations-by-type';

export type GuildKillStatsResponseDtoOutputMemberRankingItem = {
  memberId: number;
  memberName: string;
  /** @nullable */
  memberAvatar: string | null;
  memberUserId: string;
  totalParticipations: number;
  participationsByType: GuildKillStatsResponseDtoOutputMemberRankingItemParticipationsByType;
};